mv data generation strategies #170
